fix(extract): ExtractDoc.Priority must be float64 (countries.priority_score is REAL)
The extract refactor added priority_score to the PendingDocs SELECT and scanned
it into an int64 field. Unit-test seed data used only integer-valued scores
(default 0), so the suite passed — but real countries have fractional scores
(e.g. 7.5), and process-extract / process-wayback-extract then crashed at the
row scan: 'converting driver.Value type float64 ("7.5") to int64'. Caught by
live-validating the new binary against a real engine DB before deploy.

Change Priority to float64 and add a regression test seeding a fractional score.

+++ control-plane/internal/worker/extract/source.go
@@ -69,7 +69,7 @@ type ExtractDoc struct {
 	WaybackTarget string
 	Attempts      int
 	CrawlJobID    int64
-	Priority      int64
+	Priority      float64 // countries.priority_score is REAL — scanning into int64 fails on non-integer scores
 	// SourceRef carries a source-specific identifier used by adapters to credit
 	// the originating body. RegionalSource stores body_code (e.g. "ECCAA");
 	// ForeignSource stores the authority code. Empty for WaybackSource.
